mejoras en movimientos

+++ base/funciones/kaf.ft_movimiento_ime.sql
@@ -160,7 +160,13 @@ BEGIN
 			fecha_mod,
 			id_usuario_mod,
 			id_responsable_depto,
-			id_persona
+			id_persona,
+            codigo,
+            id_deposito,
+            id_depto_dest,
+            id_deposito_dest,
+            id_funcionario_dest,
+            id_movimiento_motivo
           	) values(
 			v_parametros.direccion,
 			v_parametros.fecha_hasta,
@@ -182,7 +188,13 @@ BEGIN
 			null,
 			null,
 			v_id_responsable_depto,
-			v_parametros.id_persona
+			v_parametros.id_persona,
+            v_parametros.codigo,
+            v_parametros.id_deposito,
+            v_parametros.id_depto_dest,
+            v_parametros.id_deposito_dest,
+            v_parametros.id_funcionario_dest,
+            v_parametros.id_movimiento_motivo
 			)RETURNING id_movimiento into v_id_movimiento;
 
             --Verifica si es depreciacion y esta en Estado Borrador, para precargar con todos los activos fijos del departamento
@@ -194,6 +206,7 @@ BEGIN
 
             if v_cod_movimiento = 'deprec' then
 
+                --Registra todos los activos del departamento que les corresponda depreciar en el periodo solicitado
                 insert into kaf.tmovimiento_af(
                     id_movimiento,
                     id_activo_fijo,
@@ -216,7 +229,29 @@ BEGIN
                 and afij.id_depto = v_parametros.id_depto
                 and ((afij.fecha_ult_dep is null and afij.fecha_ini_dep < v_parametros.fecha_hasta) or (afij.fecha_ult_dep < v_parametros.fecha_hasta));
 
-
+            elsif v_cod_movimiento = 'transf' then
+                --Registra todos los activos del funcionario origen
+                insert into kaf.tmovimiento_af(
+                    id_movimiento,
+                    id_activo_fijo,
+                    id_cat_estado_fun,
+                    estado_reg,
+                    fecha_reg,
+                    id_usuario_reg,
+                    fecha_mod
+                )
+                select
+                v_id_movimiento,
+                afij.id_activo_fijo,
+                afij.id_cat_estado_fun,
+                'activo',
+                now(),
+                p_id_usuario,
+                null
+                from kaf.tactivo_fijo afij
+                where afij.id_funcionario = v_parametros.id_funcionario
+                and afij.estado = 'alta'
+                and afij.en_deposito = 'no';
             end if;
 			
 			--Definicion de la respuesta
@@ -263,68 +298,114 @@ BEGIN
         		where id_movimiento=v_parametros.id_movimiento;
         	end if;
 
-            --Sentencia de la modificacion
-			update kaf.tmovimiento set
-			direccion = v_parametros.direccion,
-			fecha_hasta = v_parametros.fecha_hasta,
-			id_cat_movimiento = v_parametros.id_cat_movimiento,
-			fecha_mov = v_parametros.fecha_mov,
-			id_depto = v_parametros.id_depto,
-			id_proceso_wf = v_parametros.id_proceso_wf,
-			id_estado_wf = v_parametros.id_estado_wf,
-			glosa = v_parametros.glosa,
-			id_funcionario = v_parametros.id_funcionario,
-			estado = v_parametros.estado,
-			id_oficina = v_parametros.id_oficina,
-			num_tramite = v_parametros.num_tramite,
-			fecha_mod = now(),
-			id_usuario_mod = p_id_usuario,
-			id_usuario_ai = v_parametros._id_usuario_ai,
-			usuario_ai = v_parametros._nombre_usuario_ai,
-			id_persona = v_parametros.id_persona
-			where id_movimiento=v_parametros.id_movimiento;
+            --Permite el cambio solamente si esta en estado Borrador
+            if v_rec.estado = 'borrador' then
 
-            --Verifica el tipo de movimiento para aplicar reglas
-            select 
-            cat.codigo
-            into v_cod_movimiento
-            from param.tcatalogo cat
-            where cat.id_catalogo = v_parametros.id_cat_movimiento;
+                --Sentencia de la modificacion
+                update kaf.tmovimiento set
+                direccion = v_parametros.direccion,
+                fecha_hasta = v_parametros.fecha_hasta,
+                id_cat_movimiento = v_parametros.id_cat_movimiento,
+                fecha_mov = v_parametros.fecha_mov,
+                id_depto = v_parametros.id_depto,
+                id_proceso_wf = v_parametros.id_proceso_wf,
+                id_estado_wf = v_parametros.id_estado_wf,
+                glosa = v_parametros.glosa,
+                id_funcionario = v_parametros.id_funcionario,
+                estado = v_parametros.estado,
+                id_oficina = v_parametros.id_oficina,
+                fecha_mod = now(),
+                id_usuario_mod = p_id_usuario,
+                id_usuario_ai = v_parametros._id_usuario_ai,
+                usuario_ai = v_parametros._nombre_usuario_ai,
+                id_persona = v_parametros.id_persona,
+                codigo=v_parametros.codigo,
+                id_deposito=v_parametros.id_deposito,
+                id_depto_dest=v_parametros.id_depto_dest,
+                id_deposito_dest=v_parametros.id_deposito_dest,
+                id_funcionario_dest=v_parametros.id_funcionario_dest,
+                id_movimiento_motivo=v_parametros.id_movimiento_motivo
+                where id_movimiento=v_parametros.id_movimiento;
 
-            if v_cod_movimiento = 'deprec' then
-                --Si se cambio de depto se borra el detalle y se lo vuelve a llenar
-                if v_rec.id_depto != v_parametros.id_depto then
+                --Verifica el tipo de movimiento para aplicar reglas
+                select 
+                cat.codigo
+                into v_cod_movimiento
+                from param.tcatalogo cat
+                where cat.id_catalogo = v_parametros.id_cat_movimiento;
 
-                    delete from kaf.tmovimiento_af
-                    where id_movimiento_af in (select id_movimiento_af from kaf.tmovimiento_af where id_movimiento = v_parametros.id_movimiento);
-                    
-                    delete from kaf.tmovimiento_af where id_movimiento = v_parametros.id_movimiento;
+                if v_cod_movimiento = 'deprec' then
+                    --Si se cambio de depto se borra el detalle y se lo vuelve a llenar
+                    if v_rec.id_depto != v_parametros.id_depto then
 
-                    insert into kaf.tmovimiento_af(
-                        id_movimiento,
-                        id_activo_fijo,
-                        id_cat_estado_fun,
-                        estado_reg,
-                        fecha_reg,
-                        id_usuario_reg,
-                        fecha_mod
-                    )
-                    select 
-                    v_parametros.id_movimiento,
-                    afij.id_activo_fijo,
-                    afij.id_cat_estado_fun,
-                    'activo',
-                    now(),
-                    p_id_usuario,
-                    null
-                    from kaf.tactivo_fijo afij
-                    where afij.estado = 'alta'
-                    and afij.id_depto = v_parametros.id_depto
-                    and ((afij.fecha_ult_dep is null and afij.fecha_ini_dep < v_parametros.fecha_hasta) or (afij.fecha_ult_dep < v_parametros.fecha_hasta));
+                        delete from kaf.tmovimiento_af_dep
+                        where id_movimiento_af in (select id_movimiento_af from kaf.tmovimiento_af where id_movimiento = v_parametros.id_movimiento);
+                        
+                        delete from kaf.tmovimiento_af where id_movimiento = v_parametros.id_movimiento;
+
+                        insert into kaf.tmovimiento_af(
+                            id_movimiento,
+                            id_activo_fijo,
+                            id_cat_estado_fun,
+                            estado_reg,
+                            fecha_reg,
+                            id_usuario_reg,
+                            fecha_mod
+                        )
+                        select 
+                        v_parametros.id_movimiento,
+                        afij.id_activo_fijo,
+                        afij.id_cat_estado_fun,
+                        'activo',
+                        now(),
+                        p_id_usuario,
+                        null
+                        from kaf.tactivo_fijo afij
+                        where afij.estado = 'alta'
+                        and afij.id_depto = v_parametros.id_depto
+                        and ((afij.fecha_ult_dep is null and afij.fecha_ini_dep < v_parametros.fecha_hasta) or (afij.fecha_ult_dep < v_parametros.fecha_hasta));
+
+                    end if;
+
+                elsif v_cod_movimiento = 'transf' then
+
+                    if v_rec.id_funcionario != v_parametros.id_funcionario then
+                        
+                        delete from kaf.tmovimiento_af where id_movimiento = v_parametros.id_movimiento;
+
+                        --Registra todos los activos del funcionario origen
+                        insert into kaf.tmovimiento_af(
+                            id_movimiento,
+                            id_activo_fijo,
+                            id_cat_estado_fun,
+                            estado_reg,
+                            fecha_reg,
+                            id_usuario_reg,
+                            fecha_mod
+                        )
+                        select
+                        v_id_movimiento,
+                        afij.id_activo_fijo,
+                        afij.id_cat_estado_fun,
+                        'activo',
+                        now(),
+                        p_id_usuario,
+                        null
+                        from kaf.tactivo_fijo afij
+                        where afij.id_funcionario = v_parametros.id_funcionario
+                        and afij.estado = 'alta'
+                        and afij.en_deposito = 'no';
+
+                    end if;
+
 
                 end if;
 
+            else
+                raise exception 'Modificacion no permitida, debe estar en Estado Borrador';
             end if;
+
+            
                
 			--Definicion de la respuesta
             v_resp = pxp.f_agrega_clave(v_resp,'mensaje','Movimiento de Activos Fijos modificado(a)'); 
@@ -351,9 +432,19 @@ BEGIN
             	raise exception 'Movimiento no encontrado';
             end if;
             
-			--Sentencia de la eliminacion
-			delete from kaf.tmovimiento
-            where id_movimiento=v_parametros.id_movimiento;
+            --Obtiene los datos actuales del movimiento
+            select *
+            into v_rec
+            from kaf.tmovimiento
+            where id_movimiento = v_parametros.id_movimiento;
+
+            if v_rec.estado = 'borrador' then
+    			--Sentencia de la eliminacion
+    			delete from kaf.tmovimiento
+                where id_movimiento=v_parametros.id_movimiento;
+            else
+                raise exception 'Eliminacion no permitida, debe estar en Estado Borrador';
+            end if;
                
             --Definicion de la respuesta
             v_resp = pxp.f_agrega_clave(v_resp,'mensaje','Movimiento de Activos Fijos eliminado(a)');
